ai_server/requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.5.0
pydantic-settings==2.1.0

# ML & Data Science
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78

# Utils
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
gitpython==3.1.42

# Dev & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
